Update backend connector to use added features (merge conflicts)
