+++ examples/VPython.py
@@ -4,9 +4,9 @@
 """
 
 import roboticstoolbox as rp
-# import time
+import time
 
-env = rp.backend.VPython()
+env = rp.backends.VPython()
 env.launch(g_col=[0.2, 0.2, 0.2], g_opc=0.3)
 # env.launch()
